+++ js/widgets/behaviors/forces/force-manybody.ts
@@ -4,8 +4,9 @@
  */
 import { forceManyBody as d3ForceManyBody } from 'd3-force-3d';
 
-import { isNumeric, makeForceNodeTemplate } from '../../../template-utils';
+import { makeForceNodeTemplate } from '../../../template-utils';
 import { IBehave, IForce } from '../../../tokens';
+import { isNumeric } from '../../../utils';
 
 import { ForceBehaviorModel } from './force';