+++ js/widgets/behaviors/forces/force-z.ts
@@ -4,8 +4,9 @@
  */
 import { forceZ as d3ZForce } from 'd3-force-3d';
 
-import { isNumeric, makeForceNodeTemplate } from '../../../template-utils';
+import { makeForceNodeTemplate } from '../../../template-utils';
 import { IBehave, IForce, TAnyForce } from '../../../tokens';
+import { isNumeric } from '../../../utils';
 
 import { ForceBehaviorModel } from './force';