+++ js/widgets/behaviors/index.ts
@@ -4,6 +4,8 @@
  */
 
 export * from './base';
+export * from './forces';
+export * from './shapes';
 export * from './graph-data';
 export * from './graph-image';
 export * from './link-color';
@@ -19,6 +21,6 @@ export * from './link-selection';
 export * from './link-width';
 export * from './node-color';
 export * from './node-label';
-export * from './node-size';
 export * from './node-selection';
-export * from './forces';
+export * from './node-shape';
+export * from './node-size';