+++ js/widgets/behaviors/shapes/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2023 ipyforcegraph contributors.
+ * Distributed under the terms of the Modified BSD License.
+ */
+
+export * from './text';
+export * from './ellipse';
+export * from './rectangle';